// unicontract-frontend/src/app/interface/anagrafica-local.interface.ts
export interface AnagraficaLocalInterface {
    sesso: string;
    nazione_nascita: string;
    comune_nascita: string;
    provincia_nascita: string;
    data_nascita: string;
    cf_coniuge: string;
    stato_civile: string;
    titolo_studio: string;
    nazione_residenza: string;
    comune_residenza: string;
    provincia_residenza: string;
    cap_residenza: string;
    indirizzo_residenza: string;
    civico_residenza: string;
    data_variazione_residenza: string;
    comune_fiscale: string;
    provincia_fiscale: string;
    cap_fiscale: string;
    indirizzo_fiscale: string;
    civico_fiscale: string;
    data_variazione_dom_fiscale: string;
    comune_comunicazioni: string;
    provincia_comunicazioni: string;
    cap_comunicazioni: string;
    indirizzo_comunicazioni: string;
    civico_comunicazioni: string;
    telefono_abitazione: string;
    telefono_cellulare: string;
    telefono_ufficio: string;
    fax: string;
    email: string;
    email_privata: string;
    cf: string,
}